Add message types and move directions
