+++ WalletWasabi.Fluent/ViewModels/ChatGPT/ChatAssistantViewModel.cs
@@ -13,6 +13,7 @@ using Microsoft.CodeAnalysis.CSharp.Scripting;
 using Microsoft.Extensions.DependencyInjection;
 using Newtonsoft.Json;
 using ReactiveUI;
+using WalletWasabi.Fluent.ViewModels.ChatGPT.Messages;
 
 namespace WalletWasabi.Fluent.ViewModels.ChatGPT;
 
@@ -23,64 +24,8 @@ public partial class ChatAssistantViewModel : ReactiveObject
 		ConfigureServices();
 	}
 
-	private string _initialDirections = """
-You are a helpful assistant named Wasabito, you are Wasabi Wallet operator.
-I will write text prompts and you will generate appropriate answers
-only in json format I have provided, do not add text before or after json message.
-
-The json format for the answers is as follows:
-{
-  "status": "",
-  "message": "",
-}
-The status and message properties are of type string.
-
-The json response "status" property value can be one of the following:
-- "command":  when wasabi C# scripting api command is the answer
-- "error": when not possible to answer or other problem
-- "message": when answer is only text message but not error
-
-When "status"="command" the "message" value can only be set to
-one of the following wasabi api C# scripting commands which will be executed as C# script:
-- public async Task<string> Send(string address, string amount, string[] labels)
-  command requires address and amount parameters
-- public async Task<string> Receive(string[] labels)
-  command requires labels array parameter
-- public async Task<string> Balance()
-  command does not require any parameters
-e.g. for Send command (other follow similar pattern):
-{
-  "status": "command",
-  "message": "await Send("valid BTC address", "valid BTC amount", "valid labels comma separated",
-}
-
-If user does not provide valid param to execute api command please set status=error and ask followup question to provide that info:
-e.g.:
-{
-  "status": "error",
-  "message": "Please provide valid address for send.",
-}
-
-If not enough info is provided to execute command please set status=error and ask user to provide missing information:
-e.g.:
-{
-  "status": "error",
-  "message": "Please provide valid address.",
-}
-
-If user ask question the answer please set status=message and ask user is in following format:
-{
-  "status": "message",
-  "message": "The address used the following format...",
-}
-
-You will always write answers only as json response.
-Do not add additional text before or after json.
-""";
-
 	private ChatViewModel _chat;
-	private CancellationTokenSource _cts;
-
+	private CancellationTokenSource? _cts;
 	[AutoNotify] private bool _isChatListVisible;
 	[AutoNotify] private string? _inputText = "";
 
@@ -158,6 +103,15 @@ Do not add additional text before or after json.
 									Main = MainViewModel.Instance
 								};
 								resultMessage = await CSharpScript.EvaluateAsync<string>(message, globals: globals);
+
+								if (resultMessage is null)
+								{
+									// TODO: "Error" message view model
+									Messages.Add(new AssistantMessageViewModel
+									{
+										Message = resultMessage
+									});
+								}
 							}
 						}
 						else if (assistantResult.Status == "error")
@@ -171,6 +125,12 @@ Do not add additional text before or after json.
 								// TODO:
 								resultMessage = message;
 							}
+
+							// TODO: "Error" message view model
+							Messages.Add(new AssistantMessageViewModel
+							{
+								Message = resultMessage
+							});
 						}
 						else if (assistantResult.Status == "message")
 						{
@@ -183,19 +143,35 @@ Do not add additional text before or after json.
 								// TODO:
 								resultMessage = message;
 							}
+
+
+							// TODO: "Message" message view model
+							Messages.Add(new AssistantMessageViewModel
+							{
+								Message = resultMessage
+							});
 						}
+					}
+					else
+					{
+						// TODO: "Error" or "Assistant" message view model
+						Messages.Add(new AssistantMessageViewModel
+						{
+							Message = resultMessage
+						});
 					}
 				}
 				catch (Exception e)
 				{
 					Console.WriteLine(e);
 					resultMessage = assistantResultString;
-				}
 
-				Messages.Add(new AssistantMessageViewModel
-				{
-					Message = resultMessage
-				});
+					// TODO: "Error" message view model
+					Messages.Add(new ErrorMessageViewModel()
+					{
+						Message = resultMessage
+					});
+				}
 			}
 		}
 		catch (Exception ex)
