+++ WalletWasabi.Fluent/ViewModels/ChatGPT/Messages/Actions/ReceiveActionMessageViewModel.cs
@@ -1,0 +1,5 @@
+namespace WalletWasabi.Fluent.ViewModels.ChatGPT.Messages.Actions;
+
+public partial class ReceiveActionMessageViewModel : MessageViewModel
+{
+}
