+++ WalletWasabi.Fluent/ViewModels/ChatGPT/Messages/AssistantMessageViewModel.cs
@@ -1,4 +1,4 @@
-namespace WalletWasabi.Fluent.ViewModels.ChatGPT;
+namespace WalletWasabi.Fluent.ViewModels.ChatGPT.Messages;
 
 public partial class AssistantMessageViewModel : MessageViewModel
 {
