+++ WalletWasabi.Fluent/ViewModels/ChatGPT/Messages/ErrorMessageViewModel.cs
@@ -1,0 +1,5 @@
+namespace WalletWasabi.Fluent.ViewModels.ChatGPT.Messages;
+
+public partial class ErrorMessageViewModel : MessageViewModel
+{
+}
