+++ WalletWasabi.Fluent/ViewModels/ChatGPT/Messages/MessageViewModel.cs
@@ -1,6 +1,6 @@
 using ReactiveUI;
 
-namespace WalletWasabi.Fluent.ViewModels.ChatGPT;
+namespace WalletWasabi.Fluent.ViewModels.ChatGPT.Messages;
 
 public partial class MessageViewModel : ReactiveObject
 {
