+++ WalletWasabi.Fluent/ViewModels/ChatGPT/Messages/UserMessageViewModel.cs
@@ -1,4 +1,4 @@
-namespace WalletWasabi.Fluent.ViewModels.ChatGPT;
+namespace WalletWasabi.Fluent.ViewModels.ChatGPT.Messages;
 
 public partial class UserMessageViewModel : MessageViewModel
 {
