+++ WalletWasabi.Fluent/Views/ChatGPT/ChatAssistant.axaml
@@ -2,20 +2,20 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:s="clr-namespace:WalletWasabi.Fluent.Views.ChatGPT"
+             xmlns:cvm="clr-namespace:WalletWasabi.Fluent.Views.ChatGPT"
              mc:Ignorable="d"
              xmlns:i="clr-namespace:Avalonia.Xaml.Interactivity;assembly=Avalonia.Xaml.Interactivity"
-             xmlns:searchBar="clr-namespace:WalletWasabi.Fluent.ViewModels.ChatGPT"
+             xmlns:chat="clr-namespace:WalletWasabi.Fluent.ViewModels.ChatGPT"
              xmlns:behaviors="clr-namespace:WalletWasabi.Fluent.Behaviors"
              x:CompileBindings="True"
-             x:DataType="searchBar:ChatAssistantViewModel"
+             x:DataType="chat:ChatAssistantViewModel"
              x:Class="WalletWasabi.Fluent.Views.ChatGPT.ChatAssistant">
   <UserControl.Resources>
     <Thickness x:Key="TextControlBorderThemeThicknessFocused">1</Thickness>
   </UserControl.Resources>
 
   <Design.DataContext>
-    <searchBar:ChatAssistantViewModel />
+    <chat:ChatAssistantViewModel />
   </Design.DataContext>
 
   <UserControl.Styles>
@@ -150,7 +150,7 @@
 
     <FlyoutBase.AttachedFlyout>
       <Flyout Placement="Bottom" ShowMode="Transient">
-        <s:ChatAssistantDropdown x:Name="SearchBarDropdown" Width="400" />
+        <cvm:ChatAssistantDropdown x:Name="SearchBarDropdown" Width="400" />
       </Flyout>
     </FlyoutBase.AttachedFlyout>
   </TextBox>
